# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "voter backend",
        "title": "Voter",
        "contact": {
            "name": "Maintainer Shi Yue",
            "email": "jsclndnz@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/motions": {
            "post": {
                "description": "Add the motion to the latest session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motion"
                ],
                "summary": "Add A Motion",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.MotionAdd"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Motion"
                        }
                    }
                }
            }
        },
        "/motions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motion"
                ],
                "summary": "Get A Motion",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Motion"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.Motion": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "abstain": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.User"
                    }
                },
                "against": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.User"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "for": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.User"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sessionID": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "main.MotionAdd": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "main.User": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}